refactor :  - 통화 enum을 관리하도록 수정  - 환율 조회 시에 한화를 고정값으로 사용하도록 수정  - 내역 엔티티( history ) 추가  - 토큰 생성시 userId( 이메일 ) 사용 하도록 추가  - 요청 금액 및 응답에 Pk 값, 금액 타입 수정  - Transfer에서 Quote 컨트롤러로 분리 수정  - Serivce의 계산식들은 다른 곳에서 알아야할 필요 없다고 생각되어 Quote 로 이동  - 토큰에서 userId( 이메일 ) 추출 메서드를 SecurityUtil에 추가  - 송금 요청 기능 추가

+++ src/main/java/com/moinchallenge/dto/request/QuoteRequest.java
@@ -13,9 +13,8 @@ import lombok.NoArgsConstructor;
 public class QuoteRequest {
     @NotNull(message = "송금 금액을 입력해야 합니다.")
     @Positive(message = "송금액은 1원 이상이어야 합니다.")
-    private Integer amount;
+    private long amount;
 
     @NotNull(message = "목표 통화를 입력해야 합니다.")
     private String targetCurrency;
-
 }
